+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Image 0.3.0
+
+This is the changelog for Image version 0.3.0 released on June 4th, 2022.  For older changelogs please consult the release tag on [GitHub](https://github.com/kipcole9/image/tags)
+
+### Enhancements
+
+* Add `Image.dominant_color/2` to return the dominant color of an image
+
+* Add `Image.histogram/1` to return a histgram of an image
+
+* Add `Image.get_pixel` to get the pixel value at a given `x` and `y` offset into an image.
+
 ## Image 0.2.0
 
 This is the changelog for Image version 0.2.0 released on May 18th, 2022.  For older changelogs please consult the release tag on [GitHub](https://github.com/kipcole9/image/tags)